WIP: Adding logic to remove and add next answer
Currently the logic is semi-flawed, as the past (reviewed) answers aren't
being recorded, so we are getting duplicates. The solution here is to use
local storage to record what has already been voted on. So the idea here
is (hasBeenReviewed == isCurrentlyOnScreen || wasAlreadyOnScreen), where
'wasAlreadyOnScreen' is going to be an array of cached ObjectId values.